add votable concern
